memcache: deal with filename-duplicates correctly
when there are two files with exactly the same metadata,
the second gets a virtual appendix. This appendix disappears
if the first file is renamed (ie. it's metadata changed).
In the memcache, however, the old path is stored.
Overwrite the old path if that's the case.
Otherwise the assertion is still there to catch unwanted calls.