plugins: add a first version of exif-plugin
It seems to work so far, it passes the tests.
It has the special feature that some metadata, here
DateTime (or Modify Date) is split into Year/Month/Day
Hour/Minutes/Seconds for the FS-Metadata, so that we
can create meaningful time-based folders.
The GPS-stuff is not implemented at all yet.